NEON optimize WIP.

+++ Project3/rpi-mmal-scalar/is_options.h
@@ -10,7 +10,7 @@ extern int invert, invert_rect;
 extern int update_target_color;
 extern int show_data;
 extern int highlight_matches;
-extern int color_threshold;
+extern short int color_threshold;
 extern int imstab_servo;
 extern int imstab_digital;
 extern const int chroma_subsample_sep;
